confession: I didn't really use helm...
to install linkerd2, because it's much easier to install and check by
using their linkerd cli (which is bundled in alexellis/arkade)

Always check arkade first! Opinionated defaults that work in your favor.

Sometimes the best and easiest way to install a project is not with
Helm. This is true for Flagger, which prefers to install via
Kustomization, since so many opinionated defaults are required (for
Istio, Linkerd2, Nginx, AWS App Mesh, ... the list goes on)

Many projects are forced to provide Helm charts even though they do not
work or fit as well, because some folks can only install software if it
comes packaged in Helm, or for other compatibility reasons.

Anyway, we used the arkade tool to install linkerd2 cli in our terminal
and linkerd2 mesh in our cluster, which does not invoke helm at all.